+++ packages/calcite-components/src/components/block/block.e2e.ts
@@ -5,6 +5,7 @@ import {
   delegatesToFloatingUiOwningComponent,
   disabled,
   focusable,
+  handlesActionMenuPlacements,
   hidden,
   reflects,
   renders,
@@ -146,6 +147,15 @@ describe("calcite-block", () => {
     );
   });
 
+  describe("handles action-menu placement and flipPlacements", () => {
+    handlesActionMenuPlacements(html`
+      <calcite-block heading="heading" description="description" placement="top">
+        <calcite-action text="test" icon="banana" slot="${SLOTS.headerMenuActions}"></calcite-action>
+        <div class="content">content</div>
+      </calcite-block>
+    `);
+  });
+
   it("has a loading state", async () => {
     const page = await newE2EPage({
       html: `
@@ -172,28 +182,6 @@ describe("calcite-block", () => {
     expect(clickSpy).toHaveReceivedEventTimes(1);
 
     expect(await page.find("calcite-block >>> calcite-scrim")).toBeTruthy();
-  });
-
-  it("sets placement and flipPlacements on internal calcite-action-menu", async () => {
-    const page = await newE2EPage();
-    page.setContent(html`
-      <calcite-block heading="heading" description="description" placement="top">
-        <calcite-action text="test" icon="banana" slot="${SLOTS.headerMenuActions}"></calcite-action>
-        <div class="content">content</div>
-      </calcite-block>
-    `);
-    await page.waitForChanges();
-
-    const flipPlacements = ["top", "bottom"];
-
-    const block = await page.find("calcite-block");
-    block.setProperty("flipPlacements", flipPlacements);
-    await page.waitForChanges();
-
-    const actionMenu = await page.find("calcite-block >>> calcite-action-menu");
-
-    expect(await actionMenu.getProperty("placement")).toBe("top");
-    expect(await actionMenu.getProperty("flipPlacements")).toEqual(flipPlacements);
   });
 
   it("can display/hide content", async () => {
