+++ packages/calcite-components/src/tests/commonTests/index.ts
@@ -5,7 +5,7 @@ export { reflects } from "./reflects";
 export { renders } from "./renders";
 export { disabled } from "./disabled";
 export { hidden } from "./hidden";
-export { floatingUIOwner, delegatesToFloatingUiOwningComponent } from "./floatingUI";
+export { floatingUIOwner, delegatesToFloatingUiOwningComponent, handlesActionMenuPlacements } from "./floatingUI";
 export { focusable } from "./focusable";
 export { formAssociated } from "./formAssociated";
 export { slots } from "./slots";
